backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
google-generativeai==0.8.3
SpeechRecognition==3.10.4
python-multipart==0.0.12
pydantic==2.9.2
